Fix solaris build break
git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@45671 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ dcerpc/demos/rpcping/main.c
@@ -209,10 +209,10 @@ main(
             goto error;
         }
         
-        printf("Statistics [%u]: %u\n", i, stats->count);
+        printf("Statistics [%u]: %u\n", (unsigned int)i, (unsigned int)stats->count);
         for (j = 0; j < stats->count; j++)
         {
-            printf("  [%u] = %u\n", j, stats->stats[j]);
+            printf("  [%u] = %u\n", (unsigned int)j, (unsigned int)stats->stats[j]);
         }
 
         sleep(global.sleep);
@@ -224,7 +224,7 @@ cleanup:
 
 error:
 
-    fprintf(stderr, "Error: %x\n", status);
+    fprintf(stderr, "Error: %x\n", (unsigned int)status);
 
     goto cleanup;
 }
